add SimulatorController, add uninstall
